tout marche

+++ src_bonus/main_bonus.c
@@ -6,7 +6,7 @@
 /*   By: picatrai <picatrai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/12 16:23:31 by picatrai          #+#    #+#             */
-/*   Updated: 2024/01/04 22:25:34 by picatrai         ###   ########.fr       */
+/*   Updated: 2024/01/09 21:31:04 by picatrai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void     ft_complete(t_data_bonus *data, char **argv)
 
     ft_putstr_fd("here_doc> ", 1);
     line = get_next_line(0);
-    if (ft_strncmp(line, argv[2], ft_strlen_gnl(argv[2])) == 1)
+    if (ft_strncmp(line, argv[2], ft_strlen_gnl(argv[2])) == 1 || ft_strlen(argv[2]) != ft_strlen(line) - 1)
     {
         ft_putstr_fd(line, data->fd1);
         free(line);
@@ -52,23 +52,96 @@ void     ft_complete(t_data_bonus *data, char **argv)
         free(line);
 }
 
+static int	ft_count(int n)
+{
+	int	count;
+
+	count = 0;
+	if (n == 0)
+		return (1);
+	if (n < 0)
+		count++;
+	while (n != 0)
+	{
+		n /= 10;
+		count++;
+	}
+	return (count);
+}
+
+char	*ft_itoa(int n)
+{
+	char	*itoa;
+	int		len;
+	long	long_n;
+
+	len = ft_count(n);
+	long_n = n;
+	itoa = (char *)malloc((len + 1) * sizeof(char));
+	if (itoa == NULL)
+		return (NULL);
+	itoa[len] = '\0';
+	if (long_n < 0)
+		long_n *= -1;
+	while (len > 0)
+	{
+		itoa[len - 1] = (long_n % 10) + '0';
+		long_n /= 10;
+		len--;
+	}
+	if (n < 0)
+	itoa[0] = '-';
+	return (itoa);
+}
+
+int     ft_get_str_here(char *str, t_data_bonus *data)
+{
+    int index;
+
+    index = 1;
+    while (index <= 999)
+    {
+        data->str_index = ft_itoa(index);
+        if (data->str_index == NULL)
+            return (0);
+        data->str_here = ft_strjoin(str, data->str_index);
+        if (data->str_here == NULL)
+            return (free(data->str_index), 0);
+        if (access(data->str_here, 0) != 0)
+            return (1);
+        else
+        {
+            free(data->str_here);
+            free(data->str_index);
+        }
+        index++;
+    }
+    return (0);
+}
+
 int    ft_here_doc_and_open(t_data_bonus *data, int argc, char **argv, int (*pipes)[2])
 {
-    if (ft_strncmp(argv[1], "here_doc", 8) == 0)
+    if (ft_strncmp(argv[1], "here_doc", 8) == 0 && ft_strlen(argv[1]) == 8)
     {
+        if (ft_get_str_here(".here_doc", data) == 0)
+            return (-1);
         data->fd1 = open(".here_doc", O_CREAT | O_RDWR | O_TRUNC, 0777);
         if (data->fd1 < 0)
-            return (ft_putstr_fd("error with infile\n", 2), -1);
+            return (ft_putstr_fd("error with infile\n", 2), free(data->str_here), free(data->str_index), -1);
         ft_complete(data, argv);
         close(data->fd1);
-        data->fd1 = open(".here_doc", O_RDONLY);
+        data->fd1 = open(".here_doc", O_RDWR, 0777);
         if (data->fd1 < 0)
         {
             unlink(".here_doc");
+            free(data->str_here);
+            free(data->str_index);
             return (ft_putstr_fd("error with infile\n", 2), -1);
         }
         unlink(".here_doc");
-        data->fd2 = open(argv[argc - 1], O_CREAT | O_WRONLY | O_TRUNC, 0777);
+        free(data->str_here);
+        free(data->str_index);
+        data->fd2 = open(argv[argc - 1], O_CREAT | O_WRONLY | O_APPEND, 0777);
         if (data->fd2 < 0)
             return (ft_putstr_fd("error with outfile\n", 2), ft_close_bonus(*data, pipes, 1), -1);
         return (1);
@@ -89,10 +162,12 @@ void    ft_child_bonus(t_data_bonus data, char **argv, char **envp, int (*pipes)
 {
     int fd_in;
     int fd_out;
-    // char buffer[10];//
-    // int size;//
+    int ecart;
 
-    data.args_cmd = ft_split(argv[data.index_cmd + 2], ' ');
+    ecart = 2;
+    if (data.here_res == 1)
+        ecart++;
+    data.args_cmd = ft_split(argv[data.index_cmd + ecart], ' ');
     if (data.args_cmd == NULL)
         exit(EXIT_FAILURE);
     data.path_cmd = ft_get_path_cmd(data.args_cmd[0], envp);
@@ -105,18 +180,8 @@ void    ft_child_bonus(t_data_bonus data, char **argv, char **envp, int (*pipes)
     ft_close_pipe_useless(data, pipes, &fd_in, &fd_out);
     dup2(fd_in, STDIN_FILENO);
     dup2(fd_out, STDOUT_FILENO);
-    // size = read(fd_in, &buffer, 10);
-    // ft_putnbr_fd(size, 2);
-    // ft_putstr_fd("d\n", 2);
-    // buffer[size] = '\0';
-    // ft_putstr_fd("g\n", 2);
-    // ft_putstr_fd(buffer, 2);
-    // free(buffer);
-    ft_putnbr_fd(fd_in, 2);
-    ft_putnbr_fd(fd_out, 2);
     close(fd_in);
     close(fd_out);
-    ft_putstr_fd("e\n", 2);
     execve(data.path_cmd, data.args_cmd, envp);
     free_2d(data.args_cmd);
     free(data.path_cmd);
@@ -137,11 +202,10 @@ void    suite(t_data_bonus data, char **argv, char **envp, int (*pipes)[2])
         else if (data.pid == 0)
         {
             ft_child_bonus(data, argv, envp, pipes);
-            //waitpid(-1, NULL, 0);
+            wait(NULL);
         }
         data.index_cmd++;
     }
-    // waitpid(data.pid, NULL, 0);
     ft_child_bonus(data, argv, envp, pipes);
 }
 
@@ -158,15 +222,10 @@ int     main(int argc, char **argv, char **envp)
     if (data.here_res == -1)
         return (0);
     else if (data.here_res == 0)
-    {
-        data.index_cmd = 0;
         data.nb_pipe = argc - 4;
-    }
     else
-    {
-        data.index_cmd = 1;
-        data.nb_pipe = argc - 4;
-    }
+        data.nb_pipe = argc - 5;
+    data.index_cmd = 0;
     i = 0;
     while (i < data.nb_pipe)
     {
